+++ R/warp.R
@@ -96,7 +96,8 @@ transform_grid_grid = function(x, target, threshold) {
 			if (is.na(threshold)) {
 				p12 = st_as_sf(as.data.frame(cc_x[1:2,]), coords = 1:2)
 				threshold = signif(st_distance(p12)[1,2])
-				message(paste("threshold set to", threshold, ": set a larger value if you see missing values where they shouldn't be"))
+				message(paste("threshold set to", format(threshold), 
+					 ": set a larger value if you see missing values where they shouldn't be"))
 			}
 			i = fnn$nn.index - 1
 			i[fnn$nn.dist > threshold] = NA